Termina la animacion.
Si se esta en la cima de la curva y se apreta c: las imagenes vuelven a la grilla y la camara baja.
Si se apreta b y no hay curva: no pasa nada(fantastico :S)

Bueno prueben si maso menos responde.
Si esto anda falta las normales.

+++ TP2/Primitivas Graficas/Curva/Curva.cpp
@@ -146,7 +146,7 @@ void Curva::BezierCubica(std::list<Vertice> ptosControl, std::list<Vertice> &pto
       ptosNormal.push_back(normal);
 
       if(!begin) {
-        longitudBezier+= sqrt(pow(result.x-last.x, 2) + pow(result.y-last.y, 2));
+        longitudBezier+= sqrtf(powf(result.x-last.x, 2) + powf(result.y-last.y, 2));
         distancia.insert( std::pair<int,Vertice>(longitudBezier*factor,last) );
       } else
         begin= !begin;
